Reconnect on error
Handle any errors that occur while listening for redis events by
attempting to reconnect to redis. This fixes a bug where a redis
failover would kill an entire service due to an unhandled exception
(redis `ConnectionError`) in the managed thread.

To allow redis (or the network) to recover from the condition that
caused the error, we use an exponential backoff when attempting to
reconnect. The backoff factor can be configured via the
`pubsub_backoff_factor` config.